Changing code due to wrong interpretation of framework possibilites
It's a wrong assumption that variables of the thing can be local. All the variables should be delcared in the Thing itself. This means that you don't need to have read and write handlers for every specific attribute: it will handle it "for its own", if you'll do thing.writeProperty().

Remaining to resolve: circular dependency. If I want to turn on the bulb, I set the state on TRUE. This means also that the brightness changes, as it should be at least at 1. If brightness would cause changes that affects other properties you can MAYBE just "observe" other properties and react to changes, breaking the "chain" formed if I must ask recursively for "if this thing changed, so change also that".